additional model configs, add bootstrap training mode

+++ python/rl_loop/make_new_model.py
@@ -23,10 +23,12 @@ def main(_):
   model = model_utils.new_model(name=f'p3achygo',
                                 model_config=FLAGS.model_config)
   model(
-      tf.convert_to_tensor(
-          np.random.random([batch_size] + model.input_planes_shape())),
-      tf.convert_to_tensor(
-          np.random.random([batch_size] + model.input_features_shape())))
+      tf.convert_to_tensor(np.random.random([batch_size] +
+                                            model.input_planes_shape()),
+                           dtype=tf.float32),
+      tf.convert_to_tensor(np.random.random([batch_size] +
+                                            model.input_features_shape()),
+                           dtype=tf.float32))
   model.summary()
   model.save(FLAGS.model_path)
